full matching rewriting compiles. UNTESTED

+++ barzer_el_trie.h
@@ -222,8 +222,8 @@ public:
 		return ( type == T_REWRITER ? boost::get<uint32_t>(id) : 0xffffffff );
 	}
 	uint32_t getId_uint32() const { return ( id.which() ==0 ? boost::get<uint32_t>(id) : 0xffffffff ); }
-	uint32_t getId_int() const { return ( id.which() ==2  ? boost::get<int>(id) : 0 ); }
-	uint32_t getId_double() const { return ( id.which() ==1  ? boost::get<double>(id) : 0.0 ); }
+	int getId_int() const { return ( id.which() ==2  ? boost::get<int>(id) : 0 ); }
+	double getId_double() const { return ( id.which() ==1  ? boost::get<double>(id) : 0.0 ); }
 
 	void setStop() { type = T_STOP; }
